splitted channel.py
