+++ raiden/channel/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+from raiden.channel.participant_state import ChannelEndState
+from raiden.channel.netting_channel import Channel, ChannelExternalState
+from raiden.channel.balance_proof import BalanceProof
+
+__all__ = (
+    'BalanceProof',
+    'Channel',
+    'ChannelEndState',
+    'ChannelExternalState',
+)
